% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biostrings_helper.R
\name{blast_to_xstringset}
\alias{blast_to_xstringset}
\title{blast_to_xstring}
\usage{
blast_to_xstringset(df_blast, ext_3 = 0, ext_5 = 0, xstringset, ...)
}
\arguments{
\item{df_blast}{a tibble presenting results from a blast search, eg. as
provided by \code{\link{query_blast}}. Minimal columns that need to be
present are: 'sseqid', 'sstart' and 'send'.}

\item{ext_3}{defines extraction limits; positive number of type count that
extends the intervall in 3' direction.}

\item{ext_5}{defines extraction limits; positive number of type count that
extends the intervall in 5' direction.}

\item{xstringset}{an object of type XStringSet (BioStrings package). It
contains the sequences that correspond to the ncbi database, that has been
searched.}
}
\description{
Extract DNA or AA sequences from a specified XStringSet on the basis of a
blast search and specify extraction margins.
}
\examples{

# generate example_xstringset and example_blast:
example_xstringset <- system.file("data-raw", "example_xstringset.fna" , package = "rentrezaddon", mustWork = TRUE) \%>\% Biostrings::readDNAStringSet(.)
example_blast

example_blast \%>\%
blast_to_xstringset(., xstringset = example_xstringset)

tibble(sseqid = "sequence_1", sstart = 10, send = 5) \%>\% blast_to_xstringset(., ext_3 = 0, ext_5 = 0, example_xstringset)
tibble(sseqid = "sequence_1", sstart = 5, send = 10) \%>\% blast_to_xstringset(., ext_3 = 0, ext_5 = 0, example_xstringset)
}
