% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biostrings_helper.R
\name{define_strand}
\alias{define_strand}
\title{define_strand}
\usage{
define_strand(sstart, send, ...)
}
\arguments{
\item{sstart}{sequence start}

\item{send}{sequence end}

\item{...}{for use with pmap()}
}
\value{
Output is a tibble with columns 'start', 'end' and 'strand' -
strand being either '+' (5' to 3') or '-' (3' to 5')
}
\description{
Helper function for \code{\link{blast_to_xstring}}
}
\examples{

define_strand(sstart = 10, send = 20)
define_strand(sstart = 20, send = 10)
define_strand(sstart = 20, send = 20)

# for rowwise application in a tibble:
example_blast \%>\% dplyr::transmute(sseqid, df_strand = purrr::pmap(., define_strand)) \%>\% tidyr::unnest(df_strand)



}
