% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogeny_helper.R
\name{extract_cyano_group}
\alias{extract_cyano_group}
\title{extract_cyano_group}
\usage{
extract_cyano_group(string)
}
\arguments{
\item{string}{single single character string with lineage information of a cyanobacterial species}
}
\description{
This function extracts the cyanobacterial group from a NCBI lineage.
}
\examples{
tax_entry <- entrez_fetch(db = "taxonomy", id = 167539, rettype = "xml", parsed = TRUE)
 map_dfr(., function(x) tibble(taxid = x$TaxId,
tax_entry_tibble <- XML::xmlToList(tax_entry) \%>\%
 parent_taxid = x$ParentTaxId,
 lineage = x$Lineage,
 name = x$ScientificName))

 # if you don't like XML just use following string
 # "cellular organisms; Bacteria; Terrabacteria group; Cyanobacteria/Melainabacteria group; Cyanobacteria; Synechococcales; Prochloraceae; Prochlorococcus; Prochlorococcus marinus; Prochlorococcus marinus subsp. marinus"
 extract_cyano_group(tax_entry_tibble$lineage)
}
