% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_esummary.R
\name{extract_esummary_assembly}
\alias{extract_esummary_assembly}
\title{extract_esummary_assembly}
\usage{
extract_esummary_assembly(esummary_record)
}
\arguments{
\item{esummary_record}{single record of an esummary object as returned by rentrez::entrez_summary(db = "assembly", id) with a single id.}
}
\description{
This function extracts rentrez esummary records from assembly database
}
\examples{
esummary_1 <- rentrez::entrez_summary(db = "assembly", id = "29508")
extract_esummary_assembly(esummary_record = esummary_1)
esummary_2 <- rentrez::entrez_summary(db = "assembly", id = c("29508", "31208"))
esummary_2 \%>\% map_dfr(., extract_esummary_assembly)
}
\seealso{
Other functions for the extraction of esummaries: 
\code{\link{extract_esummary_nuccore}()},
\code{\link{extract_esummary_protein}()}
}
\concept{functions for the extraction of esummaries}
\keyword{assembly}
\keyword{esummary}
