% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_esummary.R
\name{extract_esummary_nuccore}
\alias{extract_esummary_nuccore}
\title{extract_esummary_nuccore}
\usage{
extract_esummary_nuccore(esummary_record)
}
\arguments{
\item{esummary_record}{single record of an esummary object as returned by rentrez::entrez_summary(db = "nuccore", id) with a single id.}
}
\description{
This function extracts rentrez esummary records from nuccore database
}
\examples{
esummary_1 <- rentrez::entrez_summary(db = "nuccore", id = "17227497")
extract_esummary_nuccore(esummary_record = esummary_1)
esummary_2 <- rentrez::entrez_summary(db = "nuccore", id = c("17227497", "33239452"))
esummary_2 \%>\% map_dfr(., extract_esummary_nuccore)
# Shorter way to say a similar thing; in this case set output of rentret::entrez_summary always to list, even for single records
esummary_2 \%>\% map_dfr(., magrittr::extract, c("uid", "caption", "title", "taxid", "slen", "biomol", "moltype", "topology", "sourcedb", "genome", "strand", "assemblyacc", "completeness"))
}
\seealso{
Other functions for the extraction of esummaries: 
\code{\link{extract_esummary_assembly}()},
\code{\link{extract_esummary_protein}()}
}
\concept{functions for the extraction of esummaries}
\keyword{esummary}
\keyword{nuccore}
