% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_esummary.R
\name{extract_esummary_protein}
\alias{extract_esummary_protein}
\title{extract_esummary_protein}
\usage{
extract_esummary_protein(esummary_record)
}
\arguments{
\item{esummary_record}{single record of an esummary object as returned by rentrez::entrez_summary(db = "protein", id) with a single id.}
}
\description{
This function extracts rentrez esummary records from protein database
}
\examples{
esummary_1 <- rentrez::entrez_summary(db = "protein", id = "119370761")
extract_esummary_protein(esummary_record = esummary_1)
esummary_2 <- rentrez::entrez_summary(db = "protein", id = c("119370761", "73919686"))
esummary_2 \%>\% map_dfr(., extract_esummary_protein)
# Alternatively, this is shorter:
esummary_2 \%>\% map_dfr(., magrittr::extract, c("uid", "caption", "accessionversion", "extra", "gi", "title", "taxid", "slen", "moltype", "sourcedb", "completeness", "organism"))
}
\seealso{
Other functions for the extraction of esummaries: 
\code{\link{extract_esummary_assembly}()},
\code{\link{extract_esummary_nuccore}()}
}
\concept{functions for the extraction of esummaries}
\keyword{esummary}
\keyword{protein}
