% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biostrings_helper.R
\name{extract_header_2}
\alias{extract_header_2}
\title{extract_header_2}
\usage{
extract_header_2(string, ...)
}
\arguments{
\item{string}{fasta header as character provided by names(xstringset) derived from uniprot/ncbi}
}
\description{
Extract id, title and maybe organism from uniprot/ncbi headers. Use as in 'extract_ncbi_header'
Should work for extract_pdbprf_headers, too - in beta.
}
\examples{

# example fasta / uniprot
example_header <- system.file("data-raw", "example_headers.faa" , package = "rentrezaddon", mustWork = TRUE) \%>\% Biostrings::readAAStringSet(.)

example_header \%>\% names() \%>\% map_dfr(., extract_header_2)
example_header \%>\% `names<-`(names(.) \%>\% map_dfr(., extract_header_2) \%$\% faa_id)
}
