% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biostrings_helper.R
\name{extract_id_from_xstringset}
\alias{extract_id_from_xstringset}
\title{extract_id_from_xstringset}
\usage{
extract_id_from_xstringset(string, ...)
}
\arguments{
\item{string}{fasta header as character vector provided by names(xstringset)}
}
\description{
Extract id from fasta header
}
\examples{
AAStringSet <- readAAStringSet(filepath = "/home/mirkko/Documents/NCBI/ncbi_sequences/groel_putative.fasta", format = "fasta")
names(AAStringSet) <- names(AAStringSet) \%>\% map_dfr(., extract_id_from_xstringset) \%>\% pull(id)
}
