% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biostrings_helper.R
\name{extract_mixed_header}
\alias{extract_mixed_header}
\title{extract_mixed_header}
\usage{
extract_mixed_header(string, ...)
}
\arguments{
\item{string}{fasta header as character vector provided by 'names(xstringset)' occasionally derived from ncbi}
}
\description{
Extract id, title and organism from (gi / gb / emb) header. Use as in 'extract_ncbi_header'
}
