% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biostrings_helper.R
\name{extract_ncbi_header}
\alias{extract_ncbi_header}
\title{extract_ncbi_header}
\usage{
extract_ncbi_header(string, ...)
}
\arguments{
\item{string}{fasta header as character vector provided by e.g. 'names(xstringset)' derived from ncbi}
}
\description{
Extract id, title and organism from ncbi header
}
\examples{
AAStringSet <- readAAStringSet(filepath = "/home/mirkko/Documents/NCBI/ncbi_sequences/groel_putative.fasta", format = "fasta")
names(AAStringSet) \%>\% map_dfr(., extract_ncbi_header)
}
