% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biostrings_helper.R
\name{extract_pdbprf_header}
\alias{extract_pdbprf_header}
\title{extract_pdbprf_header}
\usage{
extract_pdbprf_header(string, ...)
}
\arguments{
\item{string}{fasta header as character vector provided by names(xstringset) occasionally derived from ncbi}
}
\description{
Extract id, title but not the organism from (pdb|'id'| and prf||) headers. Use as in 'extract_ncbi_header'
}
