% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biostrings_helper.R
\name{extract_subseq}
\alias{extract_subseq}
\title{extract_subseq}
\usage{
extract_subseq(sseqid, start_extr, end_extr, strand, xstringset, ...)
}
\arguments{
\item{start_extr}{a positive integer}

\item{end_extr}{a positive integer > start_extr}

\item{strand}{either '+' (5' to 3') or '-' (3' to 5')}

\item{xstringset}{object of type XStringSet (BioStrings package)}

\item{...}{for use with pmap()}
}
\description{
Helper function for \code{\link{blast_to_xstring}}
}
\details{
Extracts a subsequence of a XString object with margins defined by start_extr
and end_extr, eg. as determined by \code{\link{define_extraction_borders}}.
If 'strand == "-" ', the subsequnece is returned as reverseComplement().
}
\examples{

# generate example_xstringset and example_blast:
example_xstringset <- system.file("data-raw", "example_xstringset.fna" , package = "rentrezaddon", mustWork = TRUE) \%>\% Biostrings::readDNAStringSet(.)
example_blast

example_blast \%>\%
dplyr::transmute(sseqid, df_strand = purrr::pmap(., define_strand)) \%>\%
tidyr::unnest(df_strand) \%>\%
dplyr::mutate(extraction = purrr::pmap(., define_extraction_borders, ext_5 = 10, ext_3 = 20, xstringset = example_xstringset)) \%>\%
unnest(extraction) \%>\%
dplyr::mutate(seq = purrr::pmap(., extract_subseq, xstringset = example_xstringset)) \%$\% seq \%>\%
Biostrings::DNAStringSetList(.) \%>\% unlist(.)
}
