% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogeny_helper.R
\name{extract_tree_label}
\alias{extract_tree_label}
\title{extract_tree_label}
\usage{
extract_tree_label(.)
}
\arguments{
\item{string}{fasta header}
}
\description{
This function extracts accessions from fasta headers
}
\examples{
extract_tree_label()
}
