% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biostrings_helper.R
\name{extract_uniprot_header}
\alias{extract_uniprot_header}
\title{extract_uniprot_header}
\usage{
extract_uniprot_header(string, ...)
}
\arguments{
\item{string}{fasta header as character vector provided by 'names(xstringset)' derived from uniprot}
}
\description{
Extract id, title and organism from uniprot header. Use as in 'extract_ncbi_header'
}
