% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_features.R
\name{featuretable_to_tibble}
\alias{featuretable_to_tibble}
\title{featuretable_to_tibble}
\usage{
featuretable_to_tibble(ft)
}
\arguments{
\item{ft}{output from 'rentrez::entrez_fetch(db = "protein", rettype = "ft", retmode = "text")'}
}
\description{
This function produces a tidy tibble of output from 'rentrez::entrez_fetch(db = "protein", rettype = "ft", retmode = "text")'
}
\examples{
# NOT WORKING (PROPERLY)!
rentrez::entrez_fetch(db = "protein", id = "BAB77899.1", rettype = "ft", retmode = "text") \%>\% featuretable_to_tibble(.)
}
