% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_entrez_links.R
\name{get_entrez_links}
\alias{get_entrez_links}
\title{get_entrez_links}
\usage{
get_entrez_links(elink_element)
}
\arguments{
\item{elink_element}{single record of an elink_element as returned by rentrez::entrez_link(db = "nuccore", dbfrom = "assembly", id) with a single id.}
}
\value{
tibble with nuccore_links (nuccore_id) and the source (refseq, insdc, other) where these links come from
}
\description{
This function extracts links from an entrez elink_element for assembly - nuccore links.
}
\details{
Links are extracted according to availability in a prioritized manner - in this order: refseq, insdc, other.
If no links are available for a given assembly, "NULL" is returned in the ouput tibble!
}
\examples{
elink <- rentrez::entrez_link(db = "nuccore", dbfrom = "assembly", id = "29508")
elink \%>\% get_entrez_links(.)
elink2 <- rentrez::entrez_link(db = "nuccore", dbfrom = "assembly", id = c("29508", "31208"), by_id = T )
elink2 \%>\% map_dfr(., get_entrez_links) \%>\% mutate(assembly_id = c("29508", "31208")) \%>\% unnest(.)
}
\keyword{elink}
