% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_ranges_helper.R
\name{get_genomic_distance}
\alias{get_genomic_distance}
\title{get_genomic_distance}
\usage{
get_genomic_distance(data, prot1, prot2)
}
\arguments{
\item{data}{tibble with minimum columns: assembly_accession, sstart, send, protein -
where sstart and send are of type doubles and refer to start and end of eg. a tblastn querry - genomic position.
protein and assembly_accession are of type character, assembly_accession ... to be continued}

\item{prot1}{name of a protein / id}

\item{prot2}{name of a protein / id}
}
\description{
Extracts genomic distances between genomic elements whose positions are known!
}
\examples{
get_genomic_distance()
protein_presence <- read_csv("/home/mirkko/Documents/NCBI/ncbi_analysis/protein_presence2.csv")
get_genomic_distance(protein_presence2, prot1 = "rbcl", prot2 = "rbcs") \%>\%
ggplot() +
geom_histogram(mapping = aes(x = distance))
}
