% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_search.R
\name{query_blastdbcmd}
\alias{query_blastdbcmd}
\title{query_blastdbcmd}
\usage{
query_blastdbcmd(db, entry, dbtype, range, strand, filepath, ...)
}
\arguments{
\item{db}{path to database}

\item{entry}{any identifier which is recognized by the local database}

\item{dbtype}{type of database - one of 'guess', 'nucl', 'prot'.}

\item{range}{optional; sequence range, given as '1-10'}

\item{strand}{optional; one of 'plus', 'minus'.}

\item{filepath}{optional; with append = TRUE}

\item{...}{to allow for additional parameters}
}
\description{
This function extracts sequences of a given range from a local database - important - with PARSED IDs!
}
\examples{
query_blastdbcmd()

query_blastdbcmd(db = "/home/mirkko/Documents/NCBI/ncbi_databases/db_genomes_parsed_id/GCF_000012525.1_ASM1252v1_protein",
entry = "WP_011242444.1",
dbtype = "prot",
filepath = "/home/mirkko/Documents/NCBI/testtesttest.faa")

c("WP_071818122.1", "WP_011242444.1") \%>\% map(., query_blastdbcmd,
db = "/home/mirkko/Documents/NCBI/ncbi_databases/db_genomes_parsed_id/GCF_000012525.1_ASM1252v1_protein",
dbtype = "prot", filepath = "/home/mirkko/Documents/NCBI/testtesttest.faa")
}
\keyword{blast}
