% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_search.R
\name{query_blastdbcmd_cdd}
\alias{query_blastdbcmd_cdd}
\title{query_blastdbcmd_cdd}
\usage{
query_blastdbcmd_cdd(db, entry, ...)
}
\arguments{
\item{db}{path to cdd_database (eg. "/home/mirkko/Documents/NCBI/ncbi_databases/cdd/Cdd_NCBI")}

\item{entry}{any identifier which is recognized by the local database, ideally obtaind by querying this local cdd database (query_blast, rpsblast)}

\item{...}{to allow for additional parameters}
}
\description{
This function extracts cdd descriptions from cdd motifs in a local cdd database.
}
\examples{
query_blastdbcmd_cdd(db = "/home/mirkko/Documents/NCBI/ncbi_databases/cdd/Cdd_NCBI", entry = "CDD:197200")
}
\keyword{blast}
