% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biostrings_helper.R
\name{writeXStringSetById}
\alias{writeXStringSetById}
\title{writeXStringSetById}
\usage{
writeXStringSetById(identifier, xstringset, path)
}
\arguments{
\item{identifier}{identifier that matches the identifier in the XStringSet object}

\item{xstringset}{XStringSet object from package 'Biostrings' (should contain the respective ids)}

\item{path}{path to preexisting directory; location of the generated files}
}
\description{
This function writes separate .fasta files from a XStringSet object to a specified path.
}
\examples{
writeXStringSetById()
}
